requirements.txt
torch
diffusers
transformers
accelerate
PIL
FastAPI
uvicorn
requests
python-dotenv
pydantic-settings
pydantic
